requirements.txt
# OpenAI + Embeddings
openai
tiktoken

# LangChain e ecossistema compatível
langchain
langchain-core
langchain-community
langchain-openai
langchain-pinecone
langchain-experimental

# Vetorização e PDF
pinecone-client
PyMuPDF
pdfplumber

# Dados e Machine Learning
pandas
numpy
scikit-learn
tabulate
joblib

# Web App
streamlit
python-dotenv

# Extras
beautifulsoup4
arxiv
wikipedia

# Notebooks
ipykernel
jupyterlab
